+++ libs/__init__.py
@@ -1,3 +1,11 @@
 # libs/__init__.py
 from .rs import Fswp
 from .keysight import *
+from .baseinstrument import BaseInstrument
+
+__all__ = \
+    [
+    'Fswp',
+    'N9020b',
+    'BaseInstrument',
+    ]